% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\docType{data}
\name{anno_double_df}
\alias{anno_double_df}
\title{Simple annotation data.frame for double precision floating point values}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 10 columns.}
\usage{
anno_double_df
}
\description{
Simple annotation data.frame for double precision floating point values
}
\keyword{datasets}
