% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-bits.R
\name{int_to_bits}
\alias{int_to_bits}
\title{Convert the lower 'nbits' of an int32 into bit representation}
\usage{
int_to_bits(int32, nbits)
}
\arguments{
\item{int32}{integer value}

\item{nbits}{number of lower bits to keep}
}
\value{
vector of length nbits
}
\description{
Convert the lower 'nbits' of an int32 into bit representation
}
