% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-chars.R
\name{plot_chars}
\alias{plot_chars}
\title{Plot bits}
\usage{
plot_chars(chars_df, annotation_df, title = NULL, base_size = 11,
  plot_theme = theme_void(base_size))
}
\arguments{
\item{chars_df}{data.frame with columns 'bit' and 'x'}

\item{annotation_df}{data.frame}

\item{title}{title}

\item{base_size}{base_size}

\item{plot_theme}{plot_theme}
}
\value{
ggplot object
}
\description{
Plot bits
}
