% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-double.R
\name{plot_compact_double}
\alias{plot_compact_double}
\title{Plot the individual bits of a double in a compact form}
\usage{
plot_compact_double(dbl, base_size = 8, legend.text.multiplier = 3,
  ...)
}
\arguments{
\item{dbl}{double}

\item{base_size}{base_size}

\item{legend.text.multiplier}{tweak size of legend text. default: 3}

\item{...}{ignored}
}
\value{
return a ggplot object
}
\description{
Plot the individual bits of a double in a compact form
}
