% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-double.R
\name{plot_double}
\alias{plot_double}
\title{Plot the individual bits of a double}
\usage{
plot_double(dbl, annotation_df = exhibitionist::anno_double_df,
  base_size = 3, ...)
}
\arguments{
\item{dbl}{double}

\item{annotation_df}{data.frame}

\item{base_size}{base_size}

\item{...}{arguments passed to \code{plot_chars}}
}
\value{
return a ggplot object
}
\description{
Plot the individual bits of a double
}
