% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-double.R
\name{plot_float}
\alias{plot_float}
\title{Plot the individual bits of a double}
\usage{
plot_float(fp, float_type = c("single", "half", "bfloat16"),
  float_bits = NULL, base_size = 8, text_y = -1, segment_y = -0.5,
  ...)
}
\arguments{
\item{fp}{floating_point stored in an integer}

\item{float_type}{one of 'single', 'half', 'bfloat16'}

\item{float_bits}{3 element vector with bit allocations for sign, exponent
and mantissa}

\item{base_size}{base_size}

\item{text_y, segment_y}{y position of the annotations for text and line segment}

\item{...}{arguments passed to \code{plot_chars}}
}
\value{
return a ggplot object
}
\description{
Plot the individual bits of a double
}
