% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-chars.R
\name{sanitise_annotation_df}
\alias{sanitise_annotation_df}
\title{Sanitise an annotation data.frame}
\usage{
sanitise_annotation_df(annotation_df, base_size = 11)
}
\arguments{
\item{annotation_df}{data.frame}

\item{base_size}{base size}
}
\value{
sanitised annotation_df data.frame
}
\description{
Sanitise an annotation data.frame
}
