% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-chars.R
\name{sanitise_chars_df}
\alias{sanitise_chars_df}
\title{Sanitise an bits data.frame}
\usage{
sanitise_chars_df(chars_df, base_size = 11)
}
\arguments{
\item{chars_df}{data.frame}

\item{base_size}{base size}
}
\value{
sanitised chars_df data.frame
}
\description{
Sanitise an bits data.frame
}
