% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-bits.R
\name{show_bits.double}
\alias{show_bits.double}
\alias{show_bits.integer}
\alias{show_bits}
\title{Show a representation of the bits in a numeric value}
\usage{
\method{show_bits}{double}(x, ...)

\method{show_bits}{integer}(x, nbits = 32, ...)

show_bits(x, ...)
}
\arguments{
\item{x}{double or integer value}

\item{...}{other arguments passed to specific implementations}

\item{nbits}{number of bits to dispaly if showing an integer}
}
\description{
Show a representation of the bits in a integer or floating point value
}
\details{
Bits are always show with the most-significant bit on the left.

For floating point numbers (i.e. 'doubles') the display is similar to the
\url{seven31}{https://github.com/ThinkR-open/seven31} package.
}
